Use ENABLE_WEBHOOKS ARG in Dockerfile for heat-operator
In an effort to reduce the computational load of our operators. We
are disabling webhooks in the service operators by default. However, they are still
required in CI for Kuttl tests to succeed. This change leverages the ENABLE_WEBHOOKS
ARG added to the Dockerfile of the Heat Operator by this PR:
https://github.com/openstack-k8s-operators/heat-operator/pull/466/files

Signed-off-by: Brendan Shephard <bshephar@redhat.com>